// src/models/IUser.ts
export interface IUser {
    id: number;
    email: string;
    name: string;
    surname: string;
    phoneNumber: string;
    roles: string[]
    building: string;
    city: string;
    street: string;
}


//      private Long id;
//     private String building;
//     private String city;
//     private String street;
//     private String email;
//     private String name;
//     private String phoneNumber;
//     private String surname;
//     private List<String> roles;